fix: spezza un ciclo di import reale e ripara i mock che mentivano (#94)
Quattro cause radice, tutte fuori dai test.

1. DIPENDENZA CIRCOLARE VERA fra RealtimePerformanceMonitor e
   RealtimeSubscriptionService. Il primo importa il secondo alla riga 10 e
   dichiara `enum ConnectionState` alla riga 13; il secondo importa quell'enum
   e lo dereferenzia alla riga 48, durante l'inizializzazione degli statici.
   Importando il monitor per primo, l'enum non esiste ancora: TypeError.

   Un commento sosteneva "Avoid circular dependency by using type-only
   import". Non era vero: un enum e' un valore, e un `import type` sarebbe
   stato cancellato dal transpiler producendo lo stesso undefined.

   NON e' un problema di test: in produzione l'esito dipende dall'ordine con
   cui il bundler valuta i moduli, cioe' da quale schermata si apre per prima.
   L'enum sta ora in types/realtime.ts, un modulo foglia - l'unica proprieta'
   che conta, perche' un file senza import non puo' partecipare a un ciclo.

2. Il mock di AsyncStorage non aveva `__esModule: true`, quindi l'interop di
   babel assegnava a `import AsyncStorage from '...'` l'intero oggetto
   `{ default: {...} }` invece del suo default: `AsyncStorage.getItem` era
   `undefined` in TUTTO il codice di produzione sotto test.

   Non si vedeva come errore perche' LocalStorageManager avvolge ogni chiamata
   in try/catch: le letture tornavano null, le scritture fallivano in
   silenzio, e dodici test asserivano su dati che nessuno aveva memorizzato.
   Un crash sarebbe stato meglio: un crash lo si cerca.
   LocalStorageManager: 12 falliti -> 1.

   Rimosso anche il SECONDO mock di AsyncStorage in jest.env.js, cieco e
   incompatibile col primo: quale dei due vincesse dipendeva dall'ordine fra
   setupFiles e setupFilesAfterEnv.

3. jest non risolveva gli import interni di react-native, perche' RN
   distribuisce varianti per piattaforma (Platform.ios.js) e nessun
   Platform.js. Qualunque test che renderizzasse un componente vero moriva su
   `Cannot find module '../Utilities/Platform'`. Aggiunta la config `haste`
   che il preset react-native imposta di suo.

4. JSX in vis-to-database-sync.test.ts, un `.ts`: 12 blocchi convertiti in
   React.createElement.

Una correzione che ho fatto e ritirato, perche' merita di restare scritta:
definire `global.window = global` sembra la risposta ovvia a
`ReferenceError: window is not defined`, e il runtime RN lo fa davvero. Ma
qui almeno due moduli decidono di essere sul web proprio da quell'assenza
(VisApiClient.ts:55, supabase.ts:18). Definirlo cambia il comportamento sotto
test di codice che nessuno stava testando, e - misurato - fa passare una
suite da verde a rossa. Il commento in jest.env.js lo spiega, cosi' nessuno
ci riprova.

  suite fallite  55 -> 55   (una carica in piu', una peggiora: vedi sopra)
  test falliti   240 -> 243
  test passati   1840 -> 1867
  test eseguiti  2082 -> 2110
  durata         245s -> 139s

(cherry picked from commit 91de59d59ac931901343bf243827f870f7339339)
